Move state into its own module.
It doesn't need to be coupled with configuration.
