+++ crates/connectors/ndc-postgres/src/lib.rs
@@ -7,3 +7,4 @@ pub mod explain;
 pub mod health;
 pub mod query;
 pub mod schema;
+pub mod state;
